+++ Maze/src/framework/game/event.h
@@ -12,10 +12,10 @@ class Event
 public:
 	EventType type;
 	int key, scancode, action, mods;
-	double x, y;
+	float x, y;
 
 	Event(int key, int scancode, int action, int mods);
-	Event(double xpos, double ypos, bool isScroll = false);
+	Event(float xpos, float ypos, bool isScroll = false);
 	~Event();
 };
